Added Collider Updates to BossBehaviours.
Boss HAS to have a Weapon now all the time in the form of a child named "Weapon", which needs at least a WeaponScript and a primitive Collider. Using multiple colliders requires them to be hardcoded into the BossBehaviours, each Collider needs a WeaponScript component aswell for ApplyDamage to function.